FEAT: add jotai to handle the sizeScreen state
